Moved Redux helpers into src/selectors
