+++ web/src/components/countryhistorycarbongraph.js
@@ -8,6 +8,8 @@ import {
   getSelectedZoneHistory,
   getZoneHistoryStartTime,
   getZoneHistoryEndTime,
+} from '../selectors';
+import {
   createSingleLayerGraphBackgroundMouseMoveHandler,
   createSingleLayerGraphBackgroundMouseOutHandler,
   createGraphLayerMouseMoveHandler,
